+++ docs/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - `--version` command line argument to display installed package version number
 - `--image_formats` command line argument to display supported image formats
+- Package configuration and now consolidated and loaded from a settings module
 ### Changed
 - Improved the output from the `--help` command line argument to make it a little clearer